Refactor canvasRoute
