+++ constants/index.js
@@ -1,0 +1,15 @@
+module.exports = {
+  OUTPUT_PATH: "./assets/images",
+  MIN_FONTSIZE: 24,
+  MAX_FONTSIZE: 38,
+  BASE_FONTFAMILY: "RobotoSlab",
+  SUB_FONTS: [
+    "Regular",
+    "Black",
+    "Bold",
+    "ExtraBold",
+    "ExtraLight",
+    "Light",
+    "Medium",
+  ],
+};
