+++ routes/canvasRoutes.js
@@ -1,0 +1,40 @@
+/**
+ * @swagger
+ * tags:
+ *   - name: Canvas
+ *     description: Operations related to canvas
+ */
+
+const express = require("express");
+const router = express.Router();
+const canvasController = require("../controllers/canvasController");
+
+/**
+ * @swagger
+ * /api/canvas:
+ *   post:
+ *     summary: Get canvas
+ *     description: Creates the canvas with words from user input
+ *     tags:
+ *       - Canvas
+ *     requestBody:
+ *       required: false
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               words:
+ *                 type: string
+ *                 description: A string of words separated by commas
+ *                 example: "Overcome, Progress, Achieve, Conquer"
+ *     responses:
+ *       '200':
+ *         description: Creates Canvas (res.type("image/png"))
+ *       '500':
+ *         description: Server error
+ */
+
+router.post("/", canvasController.getCanvas);
+
+module.exports = router;
