+++ server.js
@@ -27,7 +27,7 @@ app.get("/", (req, res) => {
 });
 
 app.use("/api/words", require("./routes/words"));
-app.use("/api/canvas", require("./routes/canvas"));
+app.use("/api/canvas", require("./routes/canvasRoutes"));
 app.use("/api/filler", require("./routes/filler"));
 
 app.listen(PORT, () => {
